Extract usefull methods to PlexusTestCaseSupport
